tests/wpt/web-platform-tests/app-history/navigate/state/updateCurrent/no-args.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
test(() => {
  assert_throws_js(TypeError, () => {
    appHistory.updateCurrent();
  }, "no args");

  assert_throws_js(TypeError, () => {
    appHistory.updateCurrent({});
  }, "empty dictionary");
}, "updateCurrent() must throw if state is not given");
</script>
